requirements.txt
requests==2.31.0
websocket-client==1.8.0